server: Add shutdown request
Add a shutdown request and corresponding peekaboo-util subcommand. To
make this possible, the frontend needs a reference to the cuckoo object
that constitutes the main loop to tell it to shut down as well. This is
a stop-gap measure until embed mode is removed so we can turn around the
architecture so that the frontend becomes the governing entity regarding
daemon lifetime.