Cortex: Add workaround for XSRF token requirement
Our usage of a retry-configured session has been observed to confuse
Cortex by learning CSRF token and even session cookies from responses
and submitting them back to the server. In keeping with the default
behaviour of the cortex4py module we configure a cookie policy in our
session that prevents any cookies from being learned from and returned
to the server.