Completely switch to asyncio
Currently Peekaboo uses classic threading for the work queue and toolbox
modules, particularly Cuckoo and Cortex job trackers. The server now
uses asyncio. This combination throws up problems such as signal
handling needing to be done by the event loop to work reliably but the
event loop getting blocked by toolbox module startup not being async.

Instead of working around these symptoms we fully switch to asyncio:

- Threads can become asyncio Tasks.

- Shutdown threading.Events can be replaced by simply cancelling the
  tasks.

- Locking mostly goes away. The Queue still needs to lock accesses to the
  duplicates logs at least where asyncio awaitables are awaited and may
  yield from the critical section modifying the log.

- sqlite (still serialised through a 1-connection QueuePool) gives no
  deadlocks/database is locked errors any more.

This state works but has a number of issues that are to be solved in
succeeding commits:

- Because of the elimination of threads we no longer scale beyond one CPU
  core.

- requests still needs replacing by aiohttp to become non-blocking.

- We only get to see any exceptions thrown in tasks created by
  asyncio.ensure_future() when retrieving their result with await on
  close_down(), i.e. on shutdown. Handing all tasks to a single
  asyncio.gather() in the daemon might remedy that.