Restore multiprocessing
This change restores multiprocessing and scaling beyond a single CPU
core by using a concurrent.futures.ThreadPoolExecutor. This allows
running blocking code in a separate thread while internally using a
queue to implement a pool - basically what our Queue was doing before
asyncio.

We add this to the sample for calculating the sha256sum and dumping
processing info and the ruleset where we use it to delegate oletools and
filetools to executors.