sample: Rework processing info dumping
Dump processing information to CouchDB instead of local files. This
allows for collecting them in a single place instead of local Peekaboo
cluster nodes and opens them up for additional analysis. We move
processing info dumping out of the sample and into the worker. This
eliminates the last common configuration element of SampleFactory which
makes the whole thing redundant. So we remove it. Instead we construct a
CouchDB client in the daemon and pass it to the workers via the queue.