server, cli: Add JSON scan endpoint
Based on our recent discovery that file name transport in multipart/mime
is broken in different ways, we try out a JSON-based endpoint. This way
we sidestep years of standard evolution and define our own interface
based on recent technology.

We put some strict schema checking on the input to formulate our
expectations and assumptions as best we can:

- the filename is a JSON string which can transport any Unicode string
  which (hopefully) should be enough for even the most strange of
  filenames.
- the sample content is base64-encoded binary data in a JSON string
- the content type and disposition are JSON strings with their values
  restricted to what the respective RFCs allow - Hah: That way we cannot
  inspect and have rules for *invalid* content type and disposition used
  in malicious mails.

As a trade-off, the file contents now need to be base64-encoded to
safely transfer binary data. This also necessitates reading in the whole
file without any possible abstractions for streaming on the client or
server side (not that we (knowingly) have any on the multipart/form-data
endpoint either).

The CLI scan-file command gets a new parameter --use-json-endpoint to
instruct it to use the new endpoint.